+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- [x] Make `write` method return a promise #4
+
+
 [0.2.0] - 2018-03-27
 --------------------
 - [x] Add `stair.connected()` attribute